Refactor models.py tests
