+++ tests/bpod_rig/calibration/liquid/test_models.py
@@ -133,25 +133,3 @@ class TestValveDataManagerJSON(unittest.TestCase):
             self.assertEqual(original_valve.amounts, loaded_valve.amounts)
             self.assertEqual(original_valve.durations, loaded_valve.durations)
             self.assertEqual(original_valve.coeffs, loaded_valve.coeffs)
-
-
-class TestSuggestDuration(unittest.TestCase):
-    def setUp(self):
-        self.valve = ValveDataClass(ValveName="Test Valve")
-        self.range_low = 2
-        self.range_high = 10
-        self.suggest_duration = lambda: utils.suggest_duration(
-            self.valve, self.range_low, self.range_high
-        )
-
-    def test_no_measures(self):
-        self.assertEqual(self.suggest_duration(), 48)
-
-    def test_one_measure(self):
-        self.valve.add_measurement(57, 7.5)
-        self.assertEqual(self.suggest_duration(), 36)
-
-    def test_two_measures(self):
-        self.valve.add_measurement(22, 2)
-        self.valve.add_measurement(66, 9.5)
-        self.assertAlmostEqual(self.suggest_duration(), 44.0, places=3)
